backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
mediapipe>=0.10.0
Pillow>=9.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
